move calculate_dimensions to database
