+++ pir/cpp/parameters.cpp
@@ -15,6 +15,7 @@
 //
 #include "pir/cpp/parameters.h"
 
+#include "pir/cpp/database.h"
 #include "seal/seal.h"
 #include "util/canonical_errors.h"
 #include "util/statusor.h"
@@ -55,20 +56,10 @@ PIRParameters CreatePIRParameters(size_t dbsize, size_t dimensions,
   PIRParameters parameters;
   parameters.set_database_size(dbsize);
   *parameters.mutable_he_parameters() = heParams;
-  for (auto& dim : CalculateDimensions(dbsize, dimensions))
+  for (auto& dim : PIRDatabase::calculate_dimensions(dbsize, dimensions))
     parameters.add_dimensions(dim);
 
   return parameters;
-}
-
-std::vector<uint32_t> CalculateDimensions(uint32_t db_size,
-                                          uint32_t num_dimensions) {
-  std::vector<uint32_t> results;
-  for (int i = num_dimensions; i > 0; --i) {
-    results.push_back(std::ceil(std::pow(db_size, 1.0 / i)));
-    db_size = std::ceil(static_cast<double>(db_size) / results.back());
-  }
-  return results;
 }
 
 }  // namespace pir
